+++ lib/constants/colors.dart
@@ -3,4 +3,5 @@ import 'package:flutter/material.dart';
 class AppColors {
   static const Color green = Color.fromRGBO(0, 153, 102, 1); //// blue #100E34
   static const Color cleargreen = Color.fromRGBO(0, 153, 102, 130);
+  static const Color selectedColor = Colors.white;
 }
