+++ lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:greenapp_flutterhack/view/loginForm.dart';
+import 'package:greenapp_flutterhack/view/homePage.dart';
+import 'package:greenapp_flutterhack/view/loginPage.dart';
 
 void main() {
   runApp(MyApp());
@@ -15,22 +16,6 @@ class MyApp extends StatelessWidget {
         theme: ThemeData(
             primarySwatch: Colors.green,
             scaffoldBackgroundColor: Color(0XFFEFF3F6)),
-        home: loginForm());
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return loginForm();
+        home: LoginPage());
   }
 }
