+++ lib/view/favoritePage.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class FavoritPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
