+++ lib/view/plantsPage.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class PlantsPage extends StatefulWidget {
+  @override
+  _PlantsPageState createState() => _PlantsPageState();
+}
+
+class _PlantsPageState extends State<PlantsPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
