+++ pubspec.yaml
@@ -27,6 +27,7 @@ dependencies:
   flutter_bloc: ^4.0.1
   equatable: ^1.2.0
   flutter_icons: ^1.1.0
+  hidden_drawer_menu: ^1.1.4
 
 
   # The following adds the Cupertino Icons font to your application.
